pyproject.toml
[project]
name = "ai-news-scraper"
version = "0.1.0"
description = "Simple ai news scraper cli tool"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "feedparser>=6.0.11",
    "google-genai>=1.18.0",
    "praw>=7.8.1",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.41",
]
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[tool.hatch.build]
include = ["readonlyai"]
